Refactor sysops (#2014)
