+++ client/internal/routemanager/systemops/systemops_nonlinux.go
@@ -10,15 +10,15 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+func (r *SysOps) AddVPNRoute(prefix netip.Prefix, intf *net.Interface) error {
+	return r.genericAddVPNRoute(prefix, intf)
+}
+
+func (r *SysOps) RemoveVPNRoute(prefix netip.Prefix, intf *net.Interface) error {
+	return r.genericRemoveVPNRoute(prefix, intf)
+}
+
 func EnableIPForwarding() error {
 	log.Infof("Enable IP forwarding is not implemented on %s", runtime.GOOS)
 	return nil
-}
-
-func AddVPNRoute(prefix netip.Prefix, intf *net.Interface) error {
-	return genericAddVPNRoute(prefix, intf)
-}
-
-func RemoveVPNRoute(prefix netip.Prefix, intf *net.Interface) error {
-	return genericRemoveVPNRoute(prefix, intf)
 }
