+++ client/internal/routemanager/systemops/systemops_unix.go
@@ -14,35 +14,33 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/netbirdio/netbird/client/internal/peer"
-	"github.com/netbirdio/netbird/iface"
 )
 
-var refCounter *ExclusionCounter
-
-func SetupRouting(initAddresses []net.IP, wgIface *iface.WGIface) (peer.BeforeAddPeerHookFunc, peer.AfterRemovePeerHookFunc, error) {
-	return setupRoutingWithRefCounter(&refCounter, initAddresses, wgIface)
+func (r *SysOps) SetupRouting(initAddresses []net.IP) (peer.BeforeAddPeerHookFunc, peer.AfterRemovePeerHookFunc, error) {
+	return r.setupRefCounter(initAddresses)
 }
 
-func CleanupRouting() error {
-	return cleanupRoutingWithRefCounter(refCounter)
+func (r *SysOps) CleanupRouting() error {
+	return r.cleanupRefCounter()
 }
 
-func addToRouteTable(prefix netip.Prefix, nexthop Nexthop) error {
-	return routeCmd("add", prefix, nexthop)
+func (r *SysOps) addToRouteTable(prefix netip.Prefix, nexthop Nexthop) error {
+	return r.routeCmd("add", prefix, nexthop)
 }
 
-func removeFromRouteTable(prefix netip.Prefix, nexthop Nexthop) error {
-	return routeCmd("delete", prefix, nexthop)
+func (r *SysOps) removeFromRouteTable(prefix netip.Prefix, nexthop Nexthop) error {
+	return r.routeCmd("delete", prefix, nexthop)
 }
 
-func routeCmd(action string, prefix netip.Prefix, nexthop Nexthop) error {
+func (r *SysOps) routeCmd(action string, prefix netip.Prefix, nexthop Nexthop) error {
 	inet := "-inet"
+	if prefix.Addr().Is6() {
+		inet = "-inet6"
+	}
+
 	network := prefix.String()
 	if prefix.IsSingleIP() {
 		network = prefix.Addr().String()
-	}
-	if prefix.Addr().Is6() {
-		inet = "-inet6"
 	}
 
 	args := []string{"-n", action, inet, network}
